Workaround for spr42761 Allegro Graph has an issue when there are uncommited changes; this is the workaround suggested by Franz.  Here's the full explanation:
As you could see from the plans, one of the differences between a Prolog
query with uncommitted triples and one without is that the former cannot
use UPI-maps whereas the latter can. Whether maps are used or not isn't
relevant but it does change the plans in question. The actual bug is
that the plan without UPI maps uses the asserted-1-db/4 functor and that
function expects to be used only with UPIs _but_ it is being used with
the future-part instead.

I will work on making a fix. In the meantime, you should be able to work
around the bug by changing your code to do something like:

[...]

This will work around the problem and also work regardless of the order
of the clauses.